+++ packages/react/__tests__/svgMock.js
@@ -1,1 +1,2 @@
+// eslint-disable-next-line react/display-name
 module.exports = () => null;