+++ packages/react/src/components/Card/CardHeader.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-export interface CardHeaderProps extends React.HTMLAttributes<HTMLDivElement> {}
+export type CardHeaderProps = React.HTMLAttributes<HTMLDivElement>;
 
 const CardHeader = ({ className, ...other }: CardHeaderProps) => (
   <div className={classNames('Card__header', className)} {...other} />