+++ packages/react/src/components/TwoColumnPanel/ColumnList.tsx
@@ -1,7 +1,8 @@
 import React, { forwardRef } from 'react';
 import classnames from 'classnames';
+import PropTypes from 'prop-types';
 
-interface ColumnListProps extends React.HTMLAttributes<HTMLDivElement> {}
+type ColumnListProps = React.HTMLAttributes<HTMLDivElement>;
 
 const ColumnList = forwardRef<HTMLDivElement, ColumnListProps>(
   ({ className, children, ...props }, ref) => {
@@ -18,5 +19,7 @@ const ColumnList = forwardRef<HTMLDivElement, ColumnListProps>(
 );
 
 ColumnList.displayName = 'ColumnList';
-
+ColumnList.propTypes = {
+  className: PropTypes.string
+};
 export default ColumnList;